The script became smarter

+++ downmodules.sh
@@ -3,13 +3,17 @@
 APPS="universal_user
 	  community
 	  mapping
+	  bc_edgv
 	  database"
 
 for app in $APPS
 do
 	echo " ++ stopping the app $app"
-	docker stop $app
-	docker rm $app
+	id="$( docker ps -a -q -f name=$app )"
+	if [ "$id" != "" ]; then
+		docker stop $app
+		docker rm $app
+	fi
 	echo " -- app $app stopped"
 done
