+++ restart_database.sh
@@ -3,9 +3,13 @@
 HOME_DIR=$(pwd)
 DB_IMAGE="database"
 
-echo " -- stopping the database"
-docker stop $DB_IMAGE
-docker rm $DB_IMAGE
+echo " ** checking if exist the container of app $DB_IMAGE"
+id="$( docker ps -a -q -f name=$DB_IMAGE )"
+if [ "$id" != "" ]; then
+	echo " -- stopping the database"
+	docker stop $DB_IMAGE
+	docker rm $DB_IMAGE
+fi
 
 echo " ++ starting the database"
 DB_ID="$( docker run -d --name $DB_IMAGE -v $HOME_DIR/data:/var/lib/postgresql/data $DB_IMAGE )"
